# Conflicts:
#	src/main/webapp/customer/bookmarks.xhtml
#	src/main/webapp/customer/borrowed.xhtml
#	src/main/webapp/media/list.xhtml

+++ src/main/java/at/qe/skeleton/services/BookmarkService.java
@@ -110,12 +110,20 @@ public class BookmarkService {
     }
 
 
+	public boolean isBookmarkedForAuthenticatedUser(final Media media) {
+		User myUser = userService.loadCurrentUser();
+		Bookmark b_check = bookmarkRepository.findFirstByUserAndMedia(myUser, media);
+
+		return b_check != null;
+	}
+
 	/**
 	 * Deletes the bookmark.
 	 *
 	 * @param bookmark the bookmark to delete
 	 */
 	public void deleteBookmark(final Bookmark bookmark) {
+		if(bookmark == null) return;
 
             bookmarkRepository.delete(bookmark);
             FacesContext context = FacesContext.getCurrentInstance();
@@ -129,9 +137,8 @@ public class BookmarkService {
 	 */
 	public void addBookmark(final Media media) {
 		User myUser = userService.loadCurrentUser();
-		Bookmark b_check = bookmarkRepository.findFirstByMedia(media);
 
-		if (b_check == null) {
+		if (!this.isBookmarkedForAuthenticatedUser(media)) {
 			Bookmark mark = new Bookmark();
 			mark.setMedia(media);
 			mark.setUser(myUser);
@@ -145,4 +152,13 @@ public class BookmarkService {
         }
 	}
 
+	public void toggleBookmark(final Media media) {
+		User myUser = userService.loadCurrentUser();
+
+		if(this.isBookmarkedForAuthenticatedUser(media)){
+			this.deleteBookmark(bookmarkRepository.findFirstByUserAndMedia(myUser, media));
+		}else{
+			this.addBookmark(media);
+		}
+	}
 }
